Implement dectection of exisitng channels in DB
If lobby channel is in DB, then that means this server process is
a recovery instance.  Do not proceed with regular InitChannels.

Channel store now takes a bool to mark whether a new channel
should be stored to channel list AND DB, or just to channel list
(in the recovery case where DB is already populated)

Deciding to keep the channel list logic in parallel with DB to
minimize changes to Talky.

Todo: recover all channels from DB instead of just hard coded
recovery of just lobby